Added autodetect feature for arduino
Now whenever new usb device is connected check runs whether it is one of the devices listed in `arduinos` array.
