+++ website/package.json
@@ -20,6 +20,7 @@
     "serialport": "^12.0.0",
     "socket.io": "^4.7.5",
     "socket.io-client": "^4.7.5",
+    "usb": "^2.14.0",
     "wake_on_lan": "^1.0.0"
   },
   "devDependencies": {
